+++ datacontract/export/exporter_factory.py
@@ -143,3 +143,7 @@ exporter_factory.register_lazy_exporter(
 exporter_factory.register_lazy_exporter(
     name=ExportFormat.spark, module_path="datacontract.export.spark_converter", class_name="SparkExporter"
 )
+
+exporter_factory.register_lazy_exporter(
+    name=ExportFormat.sqlalchemy, module_path="datacontract.export.sqlalchemy_converter", class_name="SQLAlchemyExporter"
+)